+++ examples/src/index.ts
@@ -1,2 +1,16 @@
 const a = 1;
+
+export class Example {
+  private x: number;
+
+  public constructor() {
+    this.x = 1;
+  }
+
+  public add(y: number) {
+    this.x += y;
+    console.log(this.x);
+  }
+}
+
 export default a;